Kore.Internal.TermLikeF: Split off from Kore.Internal.TermLike
