+++ kore/test/Test/Kore/Attribute/Pattern/Defined.hs
@@ -122,7 +122,7 @@ test_instance_Synthetic =
     expect
         :: GHC.HasCallStack
         => Defined
-        -> TermLikeF Variable Defined
+        -> TermLikeF (TermLike Concrete) Variable Defined
         -> TestTree
     expect x
       | isDefined x = is
