+++ kore/test/Test/Kore/Comparators.hs
@@ -268,7 +268,10 @@ instance
     , Show child
     , EqualWithExplanation variable
     , Show variable
-    ) => EqualWithExplanation (TermLikeF variable child)
+    , EqualWithExplanation key
+    , Eq key
+    , Show key
+    ) => EqualWithExplanation (TermLikeF key variable child)
   where
     compareWithExplanation = sumCompareWithExplanation
     printWithExplanation = show
@@ -2203,11 +2206,14 @@ instance
     ( EqualWithExplanation child
     , Eq child
     , Show child
+    , EqualWithExplanation key
+    , Eq key
+    , Show key
     , EqualWithExplanation variable
     , Eq variable
     , Show variable
     )
-    => SumEqualWithExplanation (TermLikeF variable child)
+    => SumEqualWithExplanation (TermLikeF key variable child)
   where
     sumConstructorPair (TermLike.AndF a1) (TermLike.AndF a2) =
         SumConstructorSameWithArguments (EqWrap "AndF" a1 a2)
